raycaster when we have a last horizontal tile hit, use it to limit the search space
Given that the raycaster moves sequentially over the FOV, we can compare the horizontal position of the last tile hit the players position.

The first couple of calls aren't efficient as they are actually look for more tiles, so perhaps some sort of restriction to only kick in on 2nd or third time we've found a tile.